+++ lib/fast_ignore.rb
@@ -3,12 +3,15 @@
 require_relative './fast_ignore/backports'
 
 require 'set'
-require_relative './fast_ignore/rule_set_builder'
-require_relative './fast_ignore/rule_builder'
+require 'strscan'
+require_relative './fast_ignore/rule_sets'
 require_relative './fast_ignore/rule_set'
+require_relative './fast_ignore/global_gitignore'
+require_relative './fast_ignore/rule_builder'
+require_relative './fast_ignore/gitignore_rule_builder'
 require_relative './fast_ignore/rule'
+require_relative './fast_ignore/unmatchable_rule'
 require_relative './fast_ignore/shebang_rule'
-require_relative './fast_ignore/fn_match_to_re'
 
 class FastIgnore
   class Error < StandardError; end
@@ -26,7 +29,7 @@ class FastIgnore
     @gitignore_enabled = gitignore
     @loaded_gitignore_files = ::Set[''] if gitignore
     @root = "#{::File.expand_path(root.to_s, Dir.pwd)}/"
-    @rule_sets = ::FastIgnore::RuleSetBuilder.build(root: @root, gitignore: gitignore, **rule_set_builder_args)
+    @rule_sets = ::FastIgnore::RuleSets.new(root: @root, gitignore: gitignore, **rule_set_builder_args)
 
     freeze
   end
@@ -34,7 +37,7 @@ class FastIgnore
   def each(&block)
     return enum_for(:each) unless block_given?
 
-    dir_pwd = Dir.pwd
+    dir_pwd = ::Dir.pwd
     root_from_pwd = @root.start_with?(dir_pwd) ? ".#{@root.delete_prefix(dir_pwd)}" : @root
 
     each_recursive(root_from_pwd, '', &block)
@@ -50,11 +53,15 @@ class FastIgnore
 
     filename = ::File.basename(relative_path)
 
-    @rule_sets.all? { |r| r.allowed_recursive?(relative_path, false, full_path, filename, content) }
+    @rule_sets.allowed_recursive?(relative_path, full_path, filename, content)
   rescue ::Errno::ENOENT, ::Errno::EACCES, ::Errno::ENOTDIR, ::Errno::ELOOP, ::Errno::ENAMETOOLONG
     false
   end
   alias_method :===, :allowed?
+
+  def to_proc
+    method(:allowed?).to_proc
+  end
 
   private
 
@@ -67,19 +74,17 @@ class FastIgnore
     paths.reverse_each(&method(:load_gitignore))
   end
 
-  def load_gitignore(parent_path, soft: true)
+  def load_gitignore(parent_path, check_exists: true)
     return if @loaded_gitignore_files.include?(parent_path)
 
-    ::FastIgnore::RuleSetBuilder.append_gitignore(
-      @rule_sets, project_root: @root, relative_path: parent_path + '.gitignore', soft: soft
-    )
+    @rule_sets.append_subdir_gitignore(relative_path: parent_path + '.gitignore', check_exists: check_exists)
+
     @loaded_gitignore_files << parent_path
   end
 
   def each_recursive(parent_full_path, parent_relative_path, &block) # rubocop:disable Metrics/MethodLength, Metrics/AbcSize
     children = ::Dir.children(parent_full_path)
-
-    load_gitignore(parent_relative_path, soft: false) if @gitignore_enabled && children.include?('.gitignore')
+    load_gitignore(parent_relative_path, check_exists: false) if @gitignore_enabled && children.include?('.gitignore')
 
     children.each do |filename|
       begin
@@ -87,7 +92,7 @@ class FastIgnore
         relative_path = parent_relative_path + filename
         dir = @follow_symlinks_method.call(full_path).directory?
 
-        next unless @rule_sets.all? { |r| r.allowed_unrecursive?(relative_path, dir, full_path, filename, nil) }
+        next unless @rule_sets.allowed_unrecursive?(relative_path, dir, full_path, filename)
 
         if dir
           each_recursive(full_path + '/', relative_path + '/', &block)