+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v0.14.0
+- significant performance improvements ~50% faster
+- add `FastIgnore#to_proc` for no good reason
+
 # v0.13.0
 - Attempt to improve documentation structure
 - Remove `gitignore: true` raising `Errno::ENOENT` if root:/.gitignore didn't exist. I can't think of a use. Now `gitignore: true` is just the default behaviour.